separate signals module + service socket communication will established only, if the module is loaded NOTE: intersections will use the signal store too.
