+++ web-app/src/app/client/modules/signals/models/signal-type-definition.model.ts
