+++ web-app/src/app/client/modules/signals/models/signal-type.model.ts
