+++ web-app/src/app/client/modules/signals/models/signal.model.ts
