+++ web-app/src/app/client/modules/signals/signal-detail/signal-detail.component.html
