+++ web-app/src/app/client/modules/signals/signal-detail/signal-detail.component.ts
@@ -5,8 +5,8 @@ import { Observable, Subscription } from 'rxjs';
 import { Signal } from '../models/signal.model';
 import { select, Store } from '@ngrx/store';
 import * as fromEep from '../store/signal.reducers';
-import { trafficLight, unknown } from '../../../shared/unicode-symbol.model';
-import { State } from '../../../app.reducers';
+import { trafficLight, unknown } from '../../../../shared/unicode-symbol.model';
+import { State } from '../../../../app.reducers';
 
 @Component({
   selector: 'app-signal-detail-component',
