+++ web-app/src/app/client/modules/signals/signal-list/signals.component.html
