+++ web-app/src/app/client/modules/signals/signal-list/signals.component.ts
@@ -3,9 +3,9 @@ import { Observable } from 'rxjs';
 import { select, Store } from '@ngrx/store';
 
 import { Signal } from '../models/signal.model';
-import * as fromRoot from '../../../app.reducers';
+import * as fromRoot from '../../../../app.reducers';
 import * as fromEep from '../store/signal.reducers';
-import { car, trafficLight } from '../../../shared/unicode-symbol.model';
+import { car, trafficLight } from '../../../../shared/unicode-symbol.model';
 import { SignalTypeDefinition } from '../models/signal-type-definition.model';
 
 @Component({
