+++ web-app/src/app/client/modules/signals/signals-routing.module.ts
@@ -5,17 +5,12 @@ import { SignalsComponent } from './signal-list/signals.component';
 import { SignalDetailComponent } from './signal-detail/signal-detail.component';
 
 const signalRoutes: Routes = [
-  {path: '', component: SignalsComponent, pathMatch: 'full'},
-  {path: ':id', component: SignalDetailComponent}
+  { path: '', component: SignalsComponent, pathMatch: 'full' },
+  { path: ':id', component: SignalDetailComponent },
 ];
 
 @NgModule({
-  imports: [
-    RouterModule.forChild(signalRoutes)
-  ],
-  exports: [
-    RouterModule
-  ]
+  imports: [RouterModule.forChild(signalRoutes)],
+  exports: [RouterModule],
 })
-export class SignalsRoutingModule {
-}
+export class SignalsRoutingModule {}
