+++ web-app/src/app/shared/ui/dashboard-sample/dashboard-sample.component.html
@@ -1,54 +1,38 @@
-<ng-container *ngIf="(connectionEstablished$ | async) === false"
-              class="container-fluid">
+<ng-container *ngIf="(connectionEstablished$ | async) === false" class="container-fluid">
   <div class="row">
     <div class="col-12">
-      <p class="mat-display-2"
-         style="margin-bottom: 24px; margin-top: 64px;">
-        EEP-Web
-      </p>
-      <p class="mat-display-1 app-mute-on-light-bg">
-        Deine Anlage. Übersichtlich. Dargestellt.
-      </p>
-      <p>Verbinde mit Server ... {{ hostName$ | async }}<br><br></p>
-      <mat-progress-spinner
-        class="my-progress"
-        color="warn"
-        mode="indeterminate">
-      </mat-progress-spinner>
+      <p class="mat-display-2" style="margin-bottom: 24px; margin-top: 64px">EEP-Web</p>
+      <p class="mat-display-1 app-mute-on-light-bg">Deine Anlage. Übersichtlich. Dargestellt.</p>
+      <p>Verbinde mit Server ... {{ hostName$ | async }}<br /><br /></p>
+      <mat-progress-spinner class="my-progress" color="warn" mode="indeterminate"> </mat-progress-spinner>
     </div>
   </div>
 </ng-container>
 
-<ng-container *ngIf="(connectionEstablished$ | async)"
-              class="container-fluid">
+<ng-container *ngIf="connectionEstablished$ | async" class="container-fluid">
   <div class="row">
     <div class="col-12">
-      <p class="mat-display-2"
-         style="margin-bottom: 24px; margin-top: 64px;">
-        EEP-Web
-      </p>
-      <p class="mat-display-1 app-mute-on-light-bg">
-        Deine Anlage. Übersichtlich. Dargestellt.
-      </p>
+      <p class="mat-display-2" style="margin-bottom: 24px; margin-top: 64px">EEP-Web</p>
+      <p class="mat-display-1 app-mute-on-light-bg">Deine Anlage. Übersichtlich. Dargestellt.</p>
     </div>
   </div>
-
 
   <h1 class="mt-5 mb-4">Schnellzugriff</h1>
   <div class="row">
     <div class="col">
-      <ng-container *ngFor="let category of navigation; let last_outer = last; let first_outer=first">
+      <ng-container *ngFor="let category of navigation; let last_outer = last; let first_outer = first">
         <ng-container *ngIf="category.name !== 'Home'">
           <ng-container *ngFor="let element of category.values; let last = last">
-            <button class="mr-3"
-                    mat-icon-button
-                    type="button"
-                    [matTooltip]="element.linkDescription"
-                    routerLink="{{ element.link }}">
-              <mat-icon
-                [matBadge]="element.badge | async"
-                matBadgePosition="above after"
-                class="app-list-icon">{{ element.icon }}</mat-icon>
+            <button
+              class="mr-3"
+              mat-icon-button
+              type="button"
+              [matTooltip]="element.linkDescription"
+              routerLink="{{ element.link }}"
+            >
+              <mat-icon [matBadge]="45" matBadgePosition="above after" class="app-list-icon">{{
+                element.icon
+              }}</mat-icon>
             </button>
           </ng-container>
         </ng-container>
@@ -56,7 +40,7 @@
     </div>
   </div>
 
-  <ng-container *ngFor="let category of navigation; let last_outer = last; let first_outer=first">
+  <ng-container *ngFor="let category of navigation; let last_outer = last; let first_outer = first">
     <ng-container *ngIf="category.name !== 'Home'">
       <h1 class="mt-5 mb-4">{{ category.name }}</h1>
 
@@ -64,9 +48,7 @@
         <ng-container *ngFor="let element of category.values; let last = last">
           <div class="col-sm-12 col-md-6 col-lg-4 d-flex">
             <mat-card class="d-flex flex-column flex-grow-1 mb-5">
-              <img *ngIf="element.image"
-                   mat-card-image
-                   src="../../../../assets/{{ element.image }}">
+              <img *ngIf="element.image" mat-card-image src="../../../../assets/{{ element.image }}" />
               <mat-card-title-group>
                 <mat-card-title>{{ element.title }}</mat-card-title>
                 <mat-card-subtitle *ngIf="element.subtitle">{{ element.subtitle }}</mat-card-subtitle>
@@ -76,9 +58,7 @@
               </mat-card-content>
               <span class="flex-grow-1"></span>
               <mat-card-actions>
-                <a mat-button
-                   color="primary"
-                   routerLink="{{ element.link }}">
+                <a mat-button color="primary" routerLink="{{ element.link }}">
                   {{ element.linkDescription }}
                 </a>
               </mat-card-actions>
@@ -86,15 +66,14 @@
           </div>
         </ng-container>
       </div>
-
     </ng-container>
   </ng-container>
 
-
   <h5 class="mt-5 mb-4 app-mute-on-light-bg">Versionen</h5>
   <p class="mat-small mat-caption app-mute-on-light-bg">
-    EEP-Web: {{ eepWebVersion$ | async }}<span
-    *ngIf="(eepVersion$ | async) !== '?'">, EEP-Lua: {{ eepLuaVersion$ | async }}, EEP: {{ eepVersion$ | async }}
-  </span>
+    EEP-Web: {{ eepWebVersion$ | async
+    }}<span *ngIf="(eepVersion$ | async) !== '?'"
+      >, EEP-Lua: {{ eepLuaVersion$ | async }}, EEP: {{ eepVersion$ | async }}
+    </span>
   </p>
 </ng-container>
